% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvec.R
\name{as_rvec}
\alias{as_rvec}
\title{Convert complete lvec to R vector}
\usage{
as_rvec(x)
}
\arguments{
\item{x}{\code{\link{lvec}} to convert.}
}
\value{
Returns an R vector of type integer, numeric, character or logical.
}
\description{
Convert complete lvec to R vector
}

