% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{chunk}
\alias{chunk}
\title{Generate a number of index ranges from a vector}
\usage{
chunk(x, chunk_size = 1e+06)
}
\arguments{
\item{x}{an object for which the index ranges should be calculated. Should 
support the \code{\link{length}} method. For example, an \code{\link{lvec}}
or a regular R vector.}

\item{chunk_size}{a numeric vector of length 1 giving the maximum length of
the chunks.}
}
\description{
The ranges have a maximum length.
}

