% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvec.R
\name{is_lvec}
\alias{is_lvec}
\title{Check if an object is of type lvec}
\usage{
is_lvec(x)
}
\arguments{
\item{x}{the object to check}
}
\value{
Returns \code{TRUE} is the object is of type \code{\link{lvec}} and
\code{FALSE} otherwise.
}
\description{
Check if an object is of type lvec
}
