% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\name{length.lvec}
\alias{length.lvec}
\alias{length<-.lvec}
\title{Get and set the length of an lvec}
\usage{
\method{length}{lvec}(x)

\method{length}{lvec}(x) <- value
}
\arguments{
\item{x}{the \code{\link{lvec}}}

\item{value}{the new length of the \code{link{lvec}}}
}
\value{
The length of the \code{\link{lvec}}.
}
\description{
Get and set the length of an lvec
}
