% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lset.R
\name{lset}
\alias{lset}
\title{Set values in an lvec}
\usage{
lset(x, index, values)
}
\arguments{
\item{x}{\code{\link{lvec}} to set values in}

\item{index}{a numeric or logical vector with indices at which the values
should be set.}

\item{value}{a vector with the new values. When shorter than the length of
the indices the values are recycled.}
}
\description{
Set values in an lvec
}
\details{
Should behave in the same way as assigning and indexing to a regular
R-vector.
}
\examples{
a <- lvec(1:10)
# set second element to 20
lset(a, 2, 20)
print(a)
# set odd elements to 20
lset(a, c(TRUE, FALSE), 20)
print(a)
# values are recycled
lset(a, 1:4, 100:101)
print(a)

}

