% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvec.R
\name{lvec_type}
\alias{lvec_type}
\title{Get the type of the lvec}
\usage{
lvec_type(x)
}
\arguments{
\item{x}{the \code{\link{lvec}} to get the type from.}
}
\value{
Returns an character vector of length 1 with one of the following values:
\code{"integer"}, \code{"numeric"}, \code{"logical"} or \code{"character"}.
}
\description{
Get the type of the lvec
}

