% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\name{order}
\alias{order}
\alias{order.default}
\alias{order.lvec}
\title{Order a lvec}
\usage{
order(x, ...)

\method{order}{default}(x, ...)

\method{order}{lvec}(x, ...)
}
\arguments{
\item{x}{lvec to sort}

\item{...}{unused.}
}
\value{
Returns the order of \code{x}. Unlike the default \code{\link{order}} 
function in R, the sort used is not stable (e.g. in case there are multiple
records with the same value in \code{x}, there relative order after sorting
is not defined).
}
\description{
Order a lvec
}
\examples{
x <- as_lvec(rnorm(10))
order(x)

}

