% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.lvec}
\alias{print.lvec}
\title{Print an lvec}
\usage{
\method{print}{lvec}(x, ...)
}
\arguments{
\item{x}{\code{\link{lvec}} to print.}

\item{...}{unused}
}
\value{
Returns \code{x} invisibly.
}
\description{
Print an lvec
}
