% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strlen.R
\name{strlen}
\alias{strlen}
\alias{strlen<-}
\title{Get and set the maximum string length of a character lvec}
\usage{
strlen(x)

strlen(x) <- value
}
\arguments{
\item{x}{a lvec of type character.}

\item{value}{the new value of the maximum string length.}
}
\description{
Get and set the maximum string length of a character lvec
}
\examples{
a <- as.lvec('123')
strlen(a)  # = 3
# Strings are truncated to strlen
lset(a, 1, '123456')
print(a)  # '123'
strln(a) <- 5
lset(a, 1, '123456')
print(a)  # '12345'

}

