% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R, R/order.R
\name{length.lvec}
\alias{length.lvec}
\alias{length<-.lvec}
\alias{order}
\alias{order.default}
\alias{order.lvec}
\title{Get and set the length of an lvec}
\usage{
\method{length}{lvec}(x)

\method{length}{lvec}(x) <- value

order(x, ...)

\method{order}{default}(x, ...)

\method{order}{lvec}(x, ...)
}
\arguments{
\item{x}{the \code{\link{lvec}}}

\item{value}{the new length of the \code{link{lvec}}}

\item{...}{unused.}

\item{x}{lvec to sort}
}
\value{
The length of the \code{\link{lvec}}.

Returns the order of \code{x}. Unlike the default \code{\link{order}} 
function in R, the sort used is not stable (e.g. in case there are multiple
records with the same value in \code{x}, there relative order after sorting
is not defined).
}
\description{
Get and set the length of an lvec

Order a lvec
}
\examples{
x <- as_lvec(rnorm(10))
order(x)

}

